Fix join game integration tests
