+++ tests/game/game_start_test.go
@@ -10,51 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func CreateGame(token string) (GameResponse, error) {
-
-	c := http.Client{}
-	url := fmt.Sprintf("%s/api/v1/games", API_URL)
-	r, err := http.NewRequest(http.MethodPost, url, nil)
-	r.Header.Set("token", token)
-	r.Header.Set("Content-Type", "application/json")
-
-	var gr GameResponse
-	resp, err := c.Do(r)
-
-	if err != nil {
-		return gr, err
-	}
-
-	dec := json.NewDecoder(resp.Body)
-	err = dec.Decode(&gr)
-
-	if gr.GameID == "" {
-		return gr, fmt.Errorf("game id is empty")
-	}
-	return gr, err
-}
-
-func JoinGame(token string, gameID string) error {
-	c := http.Client{}
-	url := fmt.Sprintf("%s/api/v1/games/%s/join", API_URL, gameID)
-	r, err := http.NewRequest(http.MethodGet, url, nil)
-	if err != nil {
-		return err
-	}
-	r.Header.Set("token", token)
-	r.Header.Set("Content-Type", "application/json")
-
-	resp, err := c.Do(r)
-	if err != nil {
-		return err
-	}
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("join game request failed")
-	}
-
-	return nil
-}
-
 func StartGame(gameID, token string) ([]entities.Player, error, int) {
 
 	c := http.Client{}
